product create request updated

+++ app/Http/Requests/Products/CreateRequest.php
@@ -31,7 +31,7 @@ class CreateRequest extends FormRequest
             'vs' => ['required', 'integer', 'max:100'],
             'dft' => ['required', 'integer', 'min:15'],
             'dry_to_touch' => ['required', 'numeric', 'min:0'],
-            'dry_to_handle' => ['required', 'integer', 'min:0'],
+            'dry_to_handle' => ['required', 'numeric', 'min:0'],
             'min_int' => ['required', 'numeric', 'min:0'],
             'max_int' => ['required', 'min:-1', 'numeric'],
             'min_temp' => ['required', 'integer', 'min:-20'],
