+++ app/Http/Requests/Products/UpdateRequest.php
@@ -33,7 +33,7 @@ class UpdateRequest extends FormRequest
             'vs' => ['required', 'integer', 'max:100'],
             'dft' => ['required', 'integer', 'min:15'],
             'dry_to_touch' => ['required', 'numeric', 'min:0'],
-            'dry_to_handle' => ['required', 'integer', 'min:0'],
+            'dry_to_handle' => ['required', 'numeric', 'min:0'],
             'min_int' => ['required', 'numeric', 'min:0'],
             'max_int' => ['required', 'numeric', 'min:-1'],
             'min_temp' => ['required', 'integer', 'min:-20'],
